src/auth/jwt.py
# src/auth/jwt.py

import os
from datetime import datetime, timedelta
from jose import JWTError, jwt

# Secret key + algorithm - use environment variables
SECRET_KEY = os.getenv("SECRET_KEY", os.getenv("JWT_SECRET", "your_secret_key_here"))
ALGORITHM = os.getenv("JWT_ALGORITHM", "HS256")
ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", "60"))


def create_access_token(data: dict, expires_delta: int = None):
    """
    Creates a JWT access token.
    """
    to_encode = data.copy()
    
    expire = datetime.utcnow() + timedelta(
        minutes=expires_delta or ACCESS_TOKEN_EXPIRE_MINUTES
    )
    to_encode.update({"exp": expire})

    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt


def decode_access_token(token: str):
    """
    Decodes and validates a JWT token.
    """
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        return payload
    except JWTError:
        return None
