requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
fastapi==0.119.1
h11==0.16.0
idna==3.11
joblib==1.5.2
numpy==2.3.4
pandas==2.3.3
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
python-jose==3.5.0
reportlab==4.4.4
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.48.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
xgboost==3.1.1
sqlalchemy==2.0.23
asyncpg==0.30.0
alembic==1.13.1
databases==0.9.0
psycopg[binary]==3.2.12
psycopg2-binary==2.9.11
email-validator==2.1.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
passlib==1.7.4
python-multipart==0.0.20
pytest-asyncio==1.3.0
httpx
Pillow==10.3.0